fix: renaming properly
