+++ lib/tricount/class/equilibrium_transaction.dart
